requirements.txt
crewai>=0.1.0
langchain>=0.1.0
pandas>=1.5.0
numpy>=1.21.0
streamlit>=1.22.0
plotly>=5.13.0
chromadb>=0.4.0
python-dotenv>=0.19.0
pytest>=7.4.0
google-cloud-aiplatform>=1.36.0
google-generativeai==0.3.2
requests>=2.31.0
scikit-learn>=1.0.0
scipy==1.12.0
statsmodels>=0.13.0
mlxtend>=0.22.0
openai>=0.27.0
matplotlib>=3.5.0
seaborn>=0.12.0
prophet>=1.1.4
pmdarima>=2.0.3 